Refactor: sharedPreference에 저장한 token들 정리 & 패키지 정리
