+++ app/src/main/java/com/depromeet/housekeeper/model/response/ApiResult.kt
@@ -5,7 +5,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import retrofit2.HttpException
 import timber.log.Timber
-import java.net.HttpURLConnection
 
 sealed class ApiResult<out T> {
     data class Success<out T>(val value: T) : ApiResult<T>()
@@ -18,8 +17,8 @@ fun <T> safeFlow(apiFunc: suspend () -> T): Flow<ApiResult<T>> = flow {
         emit(ApiResult.Success(apiFunc.invoke()))
     }
     catch (e: HttpException) {
-        if (e.code() == HttpURLConnection.HTTP_UNAUTHORIZED || e.code() == NETWORK_ERROR) {
-            // todo 여기서 apiResult Loading return하고 BaseViewModel에서 로딩중 화면 보여주게?
+        if (e.code() == NETWORK_ERROR) {
+            // todo 여기서 apiResult Loading으로 만들고 로딩중 화면 보여주게하려면 모든 네트워크 통신을 하나의 함수(baseRequest)에서 만들어서 해야할 듯
             Timber.e("auth ${e.message}")
         }
         else emit(ApiResult.Error(code = e.code(), message = e.stackTraceToString()))
